[YuvKA.View] Fix #20
Almost no view triggered the OutputWindowViewmodel's CloseWindow()
method, so closed windows were never removed from the program's
list of open windows. The only view that did this correctly was the
VideoOutputView, which happens to be the only view belonging to
nodes with any outputs.

Also fix whitespace (tabs/spaces) in XAML