[OverlayTypes] Fix MovementVectorsOverlay modifying its input
Since Frame's copy constructor only makes a shallow copy,
the code for drawing the movement vectors into the given frame
operated on the input object, thus modifying the output of another
node. This was fixed by letting MovementVectorsOverlay construct a
deep copy of its input. This however is extremely slow.